Generate types for GNOME Data Access
In addition to GLib.List presented like a TypeScript array, needed to do the same for GLib.SList, which besides Gda other libraries appear to use as well, so types for Gio, Gtk etc got some changes. New definitions for JavaScriptCore are not related to this, they are from the latest Ubuntu 18.04 package.