Generate notify signal methods the same way as other signal methods
Notify signals are just signals with a specific callback signature.
Using the normal signal method generator make the method signature
reflect the difference between Gjs vs. node-gtk (e.g. the absence of
$obj and connect_after()).

With this commit, the callback signature of connect() method in node-gtk
changes to (...args: any[]) => void (which makes the signature
consistent with on() method). I've tested what node-gtk provides in the
callback, but what I got is an object called GParam with no useful
property. So, leave it untyped for now.