Fix logic around property's readonly & construct-only
First, fix a reversal of logic: output a "readonly" property when the
property _is_ readonly, and mark a property as readonly when a property
_is not_ writable.

Then, handle the construct-only property correctly: when a property is
construct-only, the property shouldn't vanish when not in construct
prop. Instead, the property should appear as a readonly.

The first case should fix GstObject's 'name' property appears as
'readonly' in subclasses (it isn't), while the second case should fix
GstPad's 'direction' property missing.